+++ src/main/java/be/polyscripts/contactmanagerapp/service/ContactService.java
@@ -1,8 +1,11 @@
 package be.polyscripts.contactmanagerapp.service;
 
 
+import be.polyscripts.contactmanagerapp.exceptions.ContactNotFoundException;
+import be.polyscripts.contactmanagerapp.model.Company;
 import be.polyscripts.contactmanagerapp.model.Contact;
 import be.polyscripts.contactmanagerapp.repo.ContactRepository;
+import be.polyscripts.contactmanagerapp.validator.ContactValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,11 +17,15 @@ import java.util.List;
 public class ContactService {
 
     private final ContactRepository contactRepository;
+    private final ContactValidator contactValidator = new ContactValidator();
 
     @Autowired
-    public ContactService(ContactRepository contactRepository) {this.contactRepository = contactRepository;}
+    public ContactService(ContactRepository contactRepository) {
+        this.contactRepository = contactRepository;
+    }
 
     public Contact addContact(Contact contact) {
+        contactValidator.validate(contact);
         return contactRepository.save(contact);
     }
 
@@ -27,9 +34,24 @@ public class ContactService {
     }
 
     public void deleteContact(Long id) {
+        if (!contactRepository.existsById(id)) {
+            throw new ContactNotFoundException(
+                    "Contact with id " + id + " does not exists");
+        }
+        Contact contact = this.findContactByID(id);
+        if (contact.getCompanies() != null && !contact.getCompanies().isEmpty()) {
+            for (Company company : contact.getCompanies()) {
+                contact.removeCompany(company);
+            }
+        }
         contactRepository.deleteById(id);
     }
 
-    public List<Contact> findAllContacts() { return contactRepository.findAll();}
+    public List<Contact> findAllContacts() {
+        return contactRepository.findAll();
+    }
 
+    public Contact findContactByID(Long id) {
+        return contactRepository.findContactById(id);
+    }
 }
